Private types

Private types

Private types in module OceanBoundaryLayerModels:

OceanBoundaryLayerModels.ForcingData — Method.

Forcing(; tdata=[0, year], forcingfields...)

Construct a Forcing specified at time points in tdata. The forcing fields, which are arrays of data corresponding to the times in tdata and are assumed to be measured/calculated/specified at the surface and z=0, are specified by keyword arguments. Their default values are 0tdata. Possible forcing inputs are

shortwave : incoming shortwave radiation

longwave : outgoing longwave radiation

latent : incoming latent heat flux

sensible : incoming latent heat flux

precip : precipitation

evap : evaporation

xstress : wind stress in the x-direction

ystress : wind stress in the y-direction

source