Home

OceanTurb.jl

How inappropriate to call this planet Earth when it is quite clearly Ocean.

Arthur C. Clark

OceanTurb.jl implements one-dimensional partial differential equations that model turbulent convection and diffusion in the ocean surface boundary layer. It's purpose is the exploration, development, and practical usage of ocean turbulence models.

In the scheme of things

Just beneath the surface of the ocean, atmospheric fluxes of energy, heat fresh water, salinity, and momentum due to wind, waves, precipitation, evaporation, heating, cooling, and radiation drive turbulence and mediate the exchange of quantities like heat, momentum, and carbon between the atmosphere and ocean interior. Models that approximate the effects of atmospheric forcing on turbulence and turbulent mixing in the upper ocean are critical components in ocean circulation models and coupled climate models.

Installation

Open julia, press ] at the Julian prompt to enter package manager mode, and type

pkg> add https://github.com/glwagner/OceanTurb.jl.git

This installs OceanTurb.jl from the master branch. The tests can then be run by typing

pkg> test OceanTurb

Use help mode by typing ? to find information about key functions:

help?> iterate!

which should give the information

OceanTurb.iterate! — Function.

iterate!(model; Δt, Nt)

Step model forward in time for Nt steps with step size Δt.

source

Step forward m by Δt with the backward Euler method.

source

Components

Solvers for various turbulence models are implemented in submodules of OceanTurb.jl. For example, our simplest model solves the 1D diffusion equation and can be used by writing

julia> using OceanTurb.Diffusion

In addition to simple diffusion we have models for

The K-Profile-Parameterization proposed by

A 'modular' K-Profile-Parameterization that allows mixing and matching the three separate components of a canonical 'KPP'-type model.

Pacanowski-Philander

Authors

Gregory L. Wagner.