Functions

Functions

Base.Libc.time — Method.

Get the current simulation time of the model.

source

OceanTurb.FluxBoundaryCondition — Method.

FluxBoundaryCondition(boundary, flux)

Constuct a boundary condition that specifies the flux of some field on a boundary. If flux is a function, its arguments must be synced with the expection of Model.

source

OceanTurb.GradientBoundaryCondition — Method.

GradientBoundaryCondition(boundary, flux)

Constuct a boundary condition that specifies the gradient of some field on a boundary. If flux is a function, its arguments must be synced with the expection of Model.

source

OceanTurb.ValueBoundaryCondition — Method.

ValueBoundaryCondition(boundary, flux)

Constuct a boundary condition that specifies the value of some field on a boundary. If flux is a function, its arguments must be synced with the expection of Model.

source

OceanTurb.ZeroFluxBoundaryConditions — Method.

Returns FieldBoundaryConditions with zero flux at top and bottom.

source

OceanTurb.arraytype — Method.

arraytype(grid::Grid)

Return the array type corresponding to data that lives on grid. Defaults to Array. New data types (for example, grids that exist on GPUs) must implement new array types.

source

OceanTurb.iter — Method.

Get the current iteration of the model.

source

OceanTurb.iterate! — Method.

iterate!(model, Δt, nt=1)

Step model forward in time by one time-step with step-size Δt.

source

OceanTurb.oncell — Method.

oncell(c, i)

Return the interpolation of c onto cell point i.

source

OceanTurb.onface — Method.

onface(c, i)

Return the interpolation of c onto face point i.

source

OceanTurb.set! — Method.

set!(solution, kwargs...)

Set the fields of a solution. For example, use

T0 = rand(4) S0(z) = exp(-z^2/10) set!(solution, T=T0, S=S0)

To set solution.T and solution.S to T0 and S0.

source

OceanTurb.set_bcs! — Method.

set_bcs!(model, fld, bcs::Tuple)

Set boundary conditions on fld as bcs = (bottom_bc, top_bc)

set_bcs!(model; kwargs...)

Set boundary conditions, where the keywords are fields of model.solution and their arguments are tuples of (bottom, top) boundary condition.

source

OceanTurb.set_bottom_bc! — Method.

Set the bottom boundary condition for fld in model.

source

OceanTurb.set_flux_bcs! — Method.

set_flux_bcs!(model, fld, bcs::Tuple)

Set flux boundary conditions on fld as bcs = (bottom_bc, top_bc)

set_flux_bcs!(model; kwargs...)

Set flux boundary conditions for model.

source

OceanTurb.set_top_bc! — Method.

Set the top boundary condition for fld in model.

source

OceanTurb.Δc — Method.

Return the cell spacing at index i.

source

OceanTurb.Δf — Method.

Return the face spacing at index i.

source

OceanTurb.∂z! — Method.

Calculate c = ∂f/∂z in the grid interior.

source

OceanTurb.∂z! — Method.

Calculate f = ∂c/∂z in the grid interior.

source

OceanTurb.∂z — Method.

∂z(a, i)

Return the discrete derivative of a at grid point i.

The derivative of a Field{Cell} is computed at face points, and the derviative of a Field{Face} is computed at cell points.

source

OceanTurb.∂z — Method.

Return the CellField ∂f/∂z, where f is a FaceField.

source

OceanTurb.∂z — Method.

Return ∂c/∂z at face index i.

source

OceanTurb.∂z — Method.

Return the FaceField ∂c/∂z, where c is a CellField.

source

OceanTurb.∂z — Method.

Return ∂c/∂z at face index i.

source