Functions

Functions

Base.Libc.time — Method.

Get the current simulation time of the model.

source

OceanTurb.arraytype — Method.

arraytype(grid::Grid)

Return the array type corresponding to data that lives on grid. Defaults to Array. New data types (for example, grids that exist on GPUs) must implement new array types.

source

OceanTurb.fill_bottom_ghost_cell! — Method.

fill_bottom_ghost_cell!(c, κ, model, bc)

Update the bottom ghost cell of c given the boundary condition bc, model, and diffusivity kappa. kappa is used only if a flux boundary condition is specified.

source

OceanTurb.fill_top_ghost_cell! — Method.

fill_top_ghost_cell!(c, κ, model, bc)

Update the top ghost cell of c given boundary condition bc, model, and diffusivity kappa

source

OceanTurb.iter — Method.

Get the current iteration of the model.

source

OceanTurb.iterate! — Method.

iterate!(model, Δt, nt=1)

Step model forward in time by one time-step with step-size Δt.

source

OceanTurb.oncell — Method.

oncell(f, i)

Return the interpolation of f onto cell point i.

source

OceanTurb.onface — Method.

onface(c, i)

Return the interpolation of c onto face point i.

source

OceanTurb.set! — Method.

set!(solution, kwargs...)

Set the fields of a solution. For example, use

T0 = rand(4) S0(z) = exp(-z^2/10) set!(solution, T=T0, S=S0)

To set solution.T and solution.S to T0 and S0.

source

OceanTurb.Δc — Method.

Return the cell spacing at index i.

source

OceanTurb.Δf — Method.

Return the face spacing at index i.

source

OceanTurb.∂z! — Method.

Calculate c = ∂f/∂z in the grid interior.

source

OceanTurb.∂z! — Method.

Calculate f = ∂c/∂z in the grid interior.

source

OceanTurb.∂z — Method.

∂z(a, i)

Return the discrete derivative of a at grid point i.

The derivative of a Field{Cell} is computed at face points, and the derviative of a Field{Face} is computed at cell points.

source

OceanTurb.∂z — Method.

Return the CellField ∂f/∂z, where f is a FaceField.

source

OceanTurb.∂z — Method.

Return ∂c/∂z at face index i.

source

OceanTurb.∂z — Method.

Return the FaceField ∂c/∂z, where c is a CellField.

source

OceanTurb.∂z — Method.

Return ∂c/∂z at face index i.

source