Functions

Functions

Base.Libc.time — Method.

Get the current simulation time of the model.

source

OceanTurb.BoundaryConditions — Function.

BoundaryConditions([T=Float64;] bottom = GradientBoundaryCondition(-zero(T)),
                                   top = FluxBoundaryCondition(-zero(T)))

Returns FieldBoundaryConditions with a bottom and top boundary condition. The type T is only relevant for the default values of bottom and top.

source

OceanTurb.DefaultBoundaryConditions — Function.

DefaultBoundaryConditions(T=Float64)

Returns default oceanic boundary conditions: a zero GradientBoundaryCondition on bottom and a zero FluxBoundaryCondition on top.

source

OceanTurb.ZeroFluxBoundaryConditions — Function.

ZeroFluxBoundaryConditions(T=Float64)

Construct FieldBoundaryConditions with a zero FluxBoundaryCondition at top and bottom.

source

OceanTurb.absolute_error — Function.

absolute_error(c, d, p=2)

Compute the absolute error between c and d with norm p, defined as

$\mathrm{abs \, error} = \left ( L^{-1} \int_{-L}^0 |c-d|^p \, \mathrm{d} z \right )^(1/p)$.

When c and d are on different grids, d is interpolated to the same grid as c.

source

OceanTurb.arraytype — Method.

arraytype(grid::Grid)

Return the array type corresponding to data that lives on grid. Defaults to Array. New data types (for example, grids that exist on GPUs) must implement new array types.

source

OceanTurb.fill_bottom_ghost_cell! — Method.

fill_bottom_ghost_cell!(c, κ, model, bc)

Update the bottom ghost cell of c given the boundary condition bc, model, and diffusivity kappa. kappa is used only if a flux boundary condition is specified.

source

OceanTurb.fill_top_ghost_cell! — Method.

fill_top_ghost_cell!(c, κ, model, bc)

Update the top ghost cell of c given boundary condition bc, model, and diffusivity kappa

source

OceanTurb.iter — Method.

Get the current iteration of the model.

source

OceanTurb.oncell — Method.

oncell(f, i)

Return the interpolation of f onto cell point i.

source

OceanTurb.oncell — Method.

onface(c, i)

Return the interpolation of the function f onto cell point i, where f is assumed to be a function of m::AbstractModel and to return a quantity at cell interfaces.

source

OceanTurb.onface — Method.

onface(c, i)

Return the interpolation of c onto face point i.

source

OceanTurb.onface — Method.

onface(c, i)

Return the interpolation of the function f onto face point i, where f is assumed to be a function of m::AbstractModel and to return a quantity at cell centers.

source

OceanTurb.prettytime — Method.

prettytime(t)

Convert a floating point value t representing an amount of time in seconds to a more human-friendly formatted string with three decimal places. Depending on the value of t the string will be formatted to show t in nanoseconds (ns), microseconds (μs), milliseconds (ms), seconds (s), minutes (min), hours (hr), or days (day).

source

OceanTurb.relative_error — Function.

relative_error(c, d, p=2)

Compute the relative error between c and d with norm p, defined as

\[\beq \mathrm{rel \, error} = \frac{\left ( int_{-L}^0 (c-d)^p \, \mathrm{d} z \right )^(1/p)} {\left ( int_{-L}^0 d^p \, \mathrm{d} z \right )^(1/p)} \eeq\]

source

OceanTurb.run_until! — Method.

run_until!(model, dt, tfinal)

Run model until tfinal with time-step dt.

source

OceanTurb.set! — Method.

set!(solution, kwargs...)

Set the fields of a solution. For example, use

T0 = rand(4) S0(z) = exp(-z^2/10) set!(solution, T=T0, S=S0)

To set solution.T and solution.S to T0 and S0.

source

OceanTurb.set_bcs! — Method.

set_bcs!(model; bcspecs...)

Set boundary conditions of model solution fields. The keyword argument name must be the name of a model solution and its value is a (bottombc, topbc) tuple.

Example

julia> set_bcs!(model, c=(FluxBoundaryCondition(-1), FluxBoundaryCondition(0)) )

source

OceanTurb.time_step! — Method.

time_step!(model; Δt, Nt)

Step model forward in time for Nt steps with step size Δt.

source

OceanTurb.time_step! — Method.

Step forward m by Δt with the backward Euler method.

source

OceanTurb.Δc — Method.

Return the cell spacing at index i.

source

OceanTurb.Δf — Method.

Return the face spacing at index i.

source

OceanTurb.∂z! — Method.

Calculate c = ∂f/∂z in the grid interior.

source

OceanTurb.∂z! — Method.

Calculate f = ∂c/∂z in the grid interior.

source

OceanTurb.∂z — Method.

∂z(a, i)

Return the discrete derivative of a at grid point i.

The derivative of a CellField is computed at face points, and the derviative of a FaceField is computed at cell points.

source

OceanTurb.∂z — Method.

Return the CellField ∂f/∂z, where f is a FaceField.

source

OceanTurb.∂z — Method.

Return ∂c/∂z at face index i.

source

OceanTurb.∂z — Method.

Return the FaceField ∂c/∂z, where c is a CellField.

source

OceanTurb.∂z — Method.

Return ∂c/∂z at face index i.

source